pyproject.toml
[tool.poetry]
name = "nonebot-plugin-antiflash"
version = "0.2.1"
description = "Anti flash pictures in groups"
authors = ["KafCoppelia <k740677208@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_antiflash"}]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tencentyun"
url = "https://mirrors.tencent.com/pypi/simple"
default = true
